refactored tls test to use helper file
